updated notebook

+++ src/lhm/utils.py
@@ -2,5 +2,5 @@
 def report_progress(iteration, total, interval=10):
     percent = (iteration / total) * 100
     progress = int(percent / interval)
-    line = '[' + '#' * progress + ' ' * ((100 / interval) - progress) + ']'
+    line = '[' + '#' * progress + ' ' * (int(100 / interval) - progress) + ']'
     print(f'{line} {percent:.1f}% completed', end='\r')
